Added display items

+++ app/src/main/java/com/peryite/familybudget/ui/contracts/BudgetItemContract.java
@@ -38,7 +38,7 @@ public interface BudgetItemContract {
 
         void onAddItemClick();
 
-        void onDeleteItemClick(Long id);
+        void onDeleteItemClick(int id);
 
         void confirmCreateItem(Item item);
